+++ src/components/DatatableV2/body/BodyRow.tsx
@@ -1,5 +1,8 @@
 import type { DatatableInstance, DatatableRow } from '../Datatable.types';
 
+import clx from 'classnames';
+
+import DetailPanel from '../panels/DetailPanel';
 import BodyCell from './BodyCell';
 
 const BodyRow = <D,>({
@@ -9,12 +12,30 @@ const BodyRow = <D,>({
   row: DatatableRow<D>;
   table: DatatableInstance<D>;
 }) => {
+  const {
+    options: { renderDetailPanel },
+  } = table;
+  const { getVisibleCells, getIsExpanded, getIsSelected } = row;
   return (
-    <tr className="ds-table-body-row ds-table-row">
-      {row.getVisibleCells().map((cell) => (
-        <BodyCell key={cell.id} cell={cell} table={table} />
-      ))}
-    </tr>
+    <>
+      <tr
+        className={clx('ds-table-body-row ds-table-row', {
+          'is-selected': getIsSelected(),
+        })}
+        style={{
+          backgroundColor: getIsSelected()
+            ? 'var(--sscds-table-color-selection)'
+            : undefined,
+        }}
+      >
+        {getVisibleCells().map((cell) => (
+          <BodyCell key={cell.id} cell={cell} table={table} />
+        ))}
+      </tr>
+      {renderDetailPanel && getIsExpanded() && (
+        <DetailPanel row={row} table={table} />
+      )}
+    </>
   );
 };