+++ src/components/DatatableV2/buttons/IconButton.tsx
@@ -1,17 +1,21 @@
 import type { ComponentProps } from 'react';
-import type { SSCIconNames } from '../../../theme/icons/icons.enums';
+import type { SSCIcons } from '../../Icon';
 
-import { forwardRef } from 'react';
+import { forwardRef, memo } from 'react';
 import styled from 'styled-components';
 
 import { getColor, getRadii } from '../../../utils';
 import { Icon } from '../../Icon';
 import { Padbox } from '../../layout';
+import { Tooltip } from '../../Tooltip';
 
 interface IconButtonProps extends ComponentProps<'button'> {
   label: string;
-  iconName: (typeof SSCIconNames)[keyof typeof SSCIconNames];
+  /* eslint-disable @typescript-eslint/ban-types */
+  iconName: SSCIcons | (string & {});
+  /* eslint-enable @typescript-eslint/ban-types */
   iconProps?: Partial<ComponentProps<typeof Icon>>;
+  isDisabled?: boolean;
 }
 
 const IconButtonRoot = styled(Padbox)`
@@ -28,6 +32,7 @@ const IconButtonRoot = styled(Padbox)`
 
   &:disabled {
     cursor: auto;
+    opacity: 0.6;
   }
 
   &:not(:disabled):hover {
@@ -35,28 +40,35 @@ const IconButtonRoot = styled(Padbox)`
     color: ${getColor('text.primary')};
   }
 
-  &:active {
+  &:not(:disabled):active {
     background: ${getColor('primary.200')};
     color: ${getColor('text.primary')};
   }
 `;
 
 const IconButton = forwardRef<HTMLButtonElement, IconButtonProps>(
-  ({ iconName, label, iconProps, ...props }, ref) => {
+  ({ iconName, label, iconProps, isDisabled, ...props }, ref) => {
     return (
-      <IconButtonRoot
-        ref={ref}
-        aria-label={label}
-        as="button"
-        paddingSize="sm"
-        paddingType="squish"
-        type="button"
-        {...props}
+      <Tooltip
+        placement="top"
+        popup={isDisabled ? undefined : label}
+        width="auto"
       >
-        <Icon name={iconName} {...iconProps} />
-      </IconButtonRoot>
+        <IconButtonRoot
+          ref={ref}
+          aria-label={label}
+          as="button"
+          disabled={isDisabled}
+          paddingSize="sm"
+          paddingType="squish"
+          type="button"
+          {...props}
+        >
+          <Icon name={iconName} {...iconProps} />
+        </IconButtonRoot>
+      </Tooltip>
     );
   },
 );
 
-export default IconButton;
+export default memo(IconButton);