+++ src/components/DatatableV2/hooks/useDisplayColumns.tsx
@@ -5,10 +5,15 @@ import type {
 
 import { useMemo } from 'react';
 
+import ExpandAllButton from '../buttons/ExpandAllButton';
+import ExpandButton from '../buttons/ExpandButton';
+import RowActionsButton from '../buttons/RowActionsButton';
 import SelectButton from '../buttons/SelectButton';
 
 export const displayColumnIds = {
+  expand: 'ssc_dt_expand',
   select: 'ssc_dt_select',
+  actions: 'ssc_dt_actions',
 };
 export const useDisplayColumns = <D,>(
   tableOptions: ParsedDatatableOptions<D>,
@@ -17,6 +22,16 @@ export const useDisplayColumns = <D,>(
     () =>
       (
         [
+          tableOptions.enableExpanding && {
+            id: displayColumnIds.expand,
+            header: '',
+            headerComponent: tableOptions.enableExpandAll
+              ? ExpandAllButton
+              : null,
+            cell: ExpandButton,
+            size: 48,
+            ...tableOptions.defaultDisplayColumn,
+          },
           tableOptions.enableRowSelection && {
             id: displayColumnIds.select,
             header: '',
@@ -25,17 +40,29 @@ export const useDisplayColumns = <D,>(
               tableOptions.enableMultiRowSelection
                 ? ({ table }) => <SelectButton table={table} isSelectAll />
                 : null,
-            cell: ({ table, row }) => <SelectButton row={row} table={table} />,
-            size: 40,
+            cell: SelectButton,
+            size: 37,
+            ...tableOptions.defaultDisplayColumn,
+          },
+          tableOptions.enableRowActions && {
+            id: displayColumnIds.actions,
+            header: '',
+            cell: ({ table, row }) => (
+              <RowActionsButton row={row} table={table} />
+            ),
+            size: 48,
             ...tableOptions.defaultDisplayColumn,
           },
         ] as DatatableColumnDef<D>[]
       ).filter(Boolean),
     [
+      tableOptions.enableExpanding,
+      tableOptions.enableExpandAll,
       tableOptions.enableRowSelection,
       tableOptions.enableSelectAll,
       tableOptions.enableMultiRowSelection,
       tableOptions.defaultDisplayColumn,
+      tableOptions.enableRowActions,
     ],
   );
 };